+++ small_text/utils/labels.py
@@ -71,12 +71,3 @@ def list_to_csr(label_list, shape, dtype=np.int64):
     data = np.ones_like(col_ind, dtype=np.int64)
 
     return csr_matrix((data, (row_ind, col_ind)), shape=shape, dtype=dtype)
-
-
-@deprecated(deprecated_in='1.1.0', to_be_removed_in='2.0.0')
-def get_flattened_unique_labels(dataset):
-    if dataset.is_multi_label:
-        labels = np.unique(dataset.y.indices)
-    else:
-        labels = np.unique(dataset.y)
-    return np.setdiff1d(labels, np.array([LABEL_UNLABELED]))